Lint + Split Selection Manipulation Methods
