+++ Sources/CodeEditTextView/TextLayoutManager/TextLayoutManager+Public.swift
@@ -15,7 +15,7 @@ extension TextLayoutManager {
     public func estimatedWidth() -> CGFloat {
         maxLineWidth
     }
-    
+
     /// Finds a text line for the given y position relative to the text view.
     ///
     /// Y values begin at the top of the view and extend down. Eg, a `0` y value would  return the first line in
@@ -27,7 +27,7 @@ extension TextLayoutManager {
     public func textLineForPosition(_ posY: CGFloat) -> TextLineStorage<TextLine>.TextLinePosition? {
         lineStorage.getLine(atPosition: posY)
     }
-    
+
     /// Finds a text line for a given text offset.
     ///
     /// This method will not do any checking for document bounds, and will simply return `nil` if the offset if negative
@@ -54,12 +54,12 @@ extension TextLayoutManager {
         guard index >= 0 && index < lineStorage.count else { return nil }
         return lineStorage.getLine(atIndex: index)
     }
-    
+
     /// Calculates the text position at the given point in the view.
     /// - Parameter point: The point to translate to text position.
     /// - Returns: The text offset in the document where the given point is laid out.
     /// - Warning: If the requested point has not been laid out or it's layout has since been invalidated by edits or
-    ///            other changes, this method will return the invalid data. For best results, ensure the text around theaskdjhlaksj
+    ///            other changes, this method will return the invalid data. For best results, ensure the text around the
     ///            point has been laid out or is visible before calling this method.
     public func textOffsetAtPoint(_ point: CGPoint) -> Int? {
         guard point.y <= estimatedHeight() else { // End position is a special case.
